Renamed Step 1 components
