+++ src/app/portal/device/_components/forms/index.tsx
@@ -1,2 +1,2 @@
 // ? This file is used to export all the forms from the user module
-export { default as BasicDetails } from "./basic-details/server";
+export { default as DeviceCategory } from "./device-category/server";
