--- src/app/portal/device/wizard/[code]/1/@basicdetails/page.tsx
@@ -1,4 +1,0 @@
-import { BasicDetails } from '~/app/portal/device/_components/forms';
-
-const FormContainer = async () => <BasicDetails />;
-export default FormContainer;
