+++ src/app/portal/device/wizard/[code]/1/@deviceCategory/error.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { XCircleIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { startTransition, useEffect } from "react";
