+++ src/app/portal/device/wizard/[code]/1/@deviceCategory/page.tsx
@@ -1,0 +1,4 @@
+import { DeviceCategory } from '~/app/portal/device/_components/forms';
+
+const FormContainer = async () => <DeviceCategory />;
+export default FormContainer;
