+++ src/app/portal/device/wizard/[code]/1/@deviceCategory/types.ts
